% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admiral.test.R
\docType{package}
\name{admiral.test-package}
\alias{admiral.test}
\alias{admiral.test-package}
\title{admiral.test: Test Data for the 'admiral' Package}
\description{
A set of Study Data Tabulation Model (SDTM) datasets from the Clinical Data Interchange Standards Consortium (CDISC) pilot project used for testing and developing Analysis Data Model (ADaM) derivations inside the 'admiral' package.
}
\author{
\strong{Maintainer}: Ben Straub \email{ben.x.straub@gsk.com}

Authors:
\itemize{
  \item Stefan Bundfuss
  \item Gordon Miller
  \item Thomas Neitmann
  \item Syed Mubasheer
}

Other contributors:
\itemize{
  \item Ross Farrugia [contributor]
  \item Shan Lee [contributor]
  \item Gopi Vegesna [contributor]
  \item Sophie Shapcott [contributor]
  \item Mahdi About [contributor]
  \item Antonio Rodríguez [contributor]
  \item Claudia Carlucci [contributor]
  \item Annie Yang [contributor]
  \item Tamara Senior [contributor]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
  \item GlaxoSmithKline LLC [copyright holder, funder]
  \item Clinical Data Interchange Standards Consortium [copyright holder]
}

}
\keyword{internal}
