% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admiral_ex}
\alias{admiral_ex}
\title{Exposure Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1182 rows and 19 columns.
}
\source{
\url{https://github.com/pharmaverse/admiral.test/blob/main/data/admiral_ex.rda}
}
\usage{
admiral_ex
}
\description{
A SDTM EX dataset updates for Ophthalmology
}
\keyword{datasets}
