% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admiral_rs}
\alias{admiral_rs}
\title{Disease Response Dataset}
\format{An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 5703 rows and 19 columns.}
\usage{
admiral_rs
}
\description{
A SDTM RS dataset simulated by Gopi Vegesna
}
\author{
Gopi Vegesna
}
\keyword{datasets}
