% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admiral_suppds}
\alias{admiral_suppds}
\title{Supplemental Disposition Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3 rows and 9 columns.
}
\usage{
admiral_suppds
}
\description{
A SDTM SUPPDS dataset from the CDISC pilot project with additional records included
}
\keyword{datasets}
