% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admiral_supptr}
\alias{admiral_supptr}
\title{Supplemental Tumor Results Dataset}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 28989 rows and 9 columns.
}
\usage{
admiral_supptr
}
\description{
A SDTM SUPPTR dataset simulated by Gopi Vegesna
}
\author{
Gopi Vegesna
}
\keyword{datasets}
