% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admiraltest.R
\docType{package}
\name{admiraltest-package}
\alias{admiraltest}
\alias{admiraltest-package}
\title{admiraltest: Test Data for the 'admiral' Package}
\description{
A set of Study Data Tabulation Model (SDTM) datasets from the Clinical
    Data Interchange Standards Consortium (CDISC) pilot project used for testing
    and developing Analysis Data Model (ADaM) derivations inside the 'admiral'
    package.
}
\author{
\strong{Maintainer}: Thomas Neitmann \email{thomas.neitmann@roche.com}

Authors:
\itemize{
  \item Stefan Bundfuss
}

Other contributors:
\itemize{
  \item Ross Farrugia [contributor]
  \item Gopi Vegesna [contributor]
  \item Sophie Shapcott [contributor]
  \item Mahdi About [contributor]
  \item Antonio Rodríguez [contributor]
  \item Claudia Carlucci [contributor]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
  \item GlaxoSmithKline LLC [copyright holder, funder]
  \item Clinical Data Interchange Standards Consortium [copyright holder]
}

}
\keyword{internal}
