% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adsl}
\alias{adsl}
\title{ADSL}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 254 rows and 48 columns.
}
\source{
\url{https://bitbucket.cdisc.org/projects/CED/repos/sdtm-adam-pilot-project/browse/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/analysis/adam/datasets/adsl.xpt}
}
\usage{
adsl
}
\description{
An ADSL dataset from the CDISC pilot project
}
\keyword{datasets}
