% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_terms.R
\name{get_smq_terms}
\alias{get_smq_terms}
\title{An example function as expected by the `get_smq_fun` parameter of
`admiral::create_query_data()`}
\usage{
get_smq_terms(smq_select, version, keep_id, temp_env)
}
\description{
An example function as expected by the `get_smq_fun` parameter of
`admiral::create_query_data()`
}
