% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admiral.test.R
\docType{package}
\name{admiral.test-package}
\alias{admiral.test}
\alias{admiral.test-package}
\title{admiral.test: Test Data for the {admiral} Package}
\description{
A set of SDTM datasets from the CDISC pilot project used for testing inside {admiral}.
}
\author{
\strong{Maintainer}: Thomas Neitmann \email{thomas.neitmann@roche.com}

Authors:
\itemize{
  \item Stefan Bundfuss
}

Other contributors:
\itemize{
  \item Clinical Data Interchange Standards Consortium [copyright holder]
}

}
\keyword{internal}
