% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admiral_mh}
\alias{admiral_mh}
\title{Medical History Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1818 rows and 26 columns.
}
\usage{
admiral_mh
}
\description{
A SDTM MH dataset from the CDISC pilot project with additional variables included
}
\keyword{datasets}
