% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pp}
\alias{pp}
\title{Pharmacokinetic Parameters Dataset}
\format{
An object of class \code{data.frame} with 2348 rows and 14 columns.
}
\usage{
pp
}
\description{
A SDTM PP dataset simulated by Antonio Rodríguez Contestí
}
\author{
Antonio Rodríguez Contestí
}
\keyword{datasets}
